backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2==2.9.11
psycopg2-binary>=2.9.0
PyJWT==2.12.1
sqlparse==0.5.5
